feat(e2e-metrics): add e2e tests for service monitor (#302)
* feat(e2e-metrics): add e2e tests for HiveMetastore service monitor

* fix(metrics): update targetPort to string format in metrics service definition

* test: fix metrcis service asserts

* feat(e2e-metrics): add wait for ZooKeeper pods readiness in metrics endpoint test

+++ internal/clustercontroller/server/svc.go
@@ -88,7 +88,7 @@ func NewRoleGroupMetricsService(
 	servicePorts := []corev1.ContainerPort{
 		{
 			Name:          zkv1alpha1.MetricsPortName,
-			ContainerPort: zkv1alpha1.MetricsPort,
+			ContainerPort: int32(zkv1alpha1.MetricsPort),
 			Protocol:      corev1.ProtocolTCP,
 		},
 	}
